backend/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ci-sdr==0.0.2
click==8.1.3
commonmark==0.9.1
ConfigArgParse==1.5.3
ctc-segmentation==1.7.4
Cython==0.29.32
datasets==2.6.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
Distance==0.1.3
docker-pycreds==0.4.0
einops==0.6.0
entrypoints==0.4
espnet==202209
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
executing==1.2.0
fast-bss-eval==0.1.3
fastapi==0.86.0
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
g2p-en==2.1.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
huggingface-hub==0.10.1
humanfriendly==10.0
idna==3.4
importlib-metadata==4.13.0
inflect==6.0.2
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
jaconv==0.3
jamo==0.4.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
kaldiio==2.17.2
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==9.0.0
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.13
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0b1
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
pooch==1.6.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyopenjtalk==0.3.0
pyparsing==3.0.9
pypinyin==0.44.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-wpe==0.0.1
pyttsx3==2.90
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyworld==0.3.2
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
responses==0.18.0
rich==12.6.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.10.1
seqeval==1.2.2
setproctitle==1.3.2
shortuuid==1.0.9
simpletransformers==0.63.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.11.0
soupsieve==2.3.2.post1
stack-data==0.6.1
starlette==0.20.4
streamlit==1.14.0
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.14.0.dev20221105+cu117
torch-complex==0.4.3
torchaudio==0.14.0.dev20221105+cu117
torchvision==0.15.0.dev20221105+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.21.0
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.12
uvicorn==0.19.0
validators==0.20.0
wandb==0.13.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
whisper @ git+https://github.com/openai/whisper.git@9f70a352f9f8630ab3aa0d06af5cb9532bd8c21d
xxhash==3.1.0
yarl==1.8.1
zipp==3.10.0
